refactor(diffs): privatize internal plugin surfaces
